+++ package.json
@@ -6,8 +6,11 @@
   "engines": {
     "node": "18.x"
   },
+  "type": "module",
   "scripts": {
-    "build": "npx uglify-js --compress -- public/scripts/dotnet-random.js public/scripts/locations.js public/scripts/items.js public/scripts/loadout.js public/scripts/logic.js public/scripts/itemPlacement.js public/scripts/modeStandard.js public/scripts/modeRecall.js public/scripts/bps-patch.js public/scripts/ips-patch.js public/scripts/sm-rando.js > public/app/dash.min.js"
+    "build:cli": "npx uglify-js --compress -- public/scripts/dotnet-random.js public/scripts/locations.js public/scripts/items.js public/scripts/loadout.js public/scripts/logic.js public/scripts/itemPlacement.js public/scripts/modeStandard.js public/scripts/modeRecall.js public/scripts/bps-patch.js public/scripts/ips-patch.js public/scripts/sm-rando.js > public/app/dash.min.js",
+    "build:web": "esbuild scripts/pages/*.js --outdir=public --bundle --minify --sourcemap --entry-names=[name].bundled",
+    "dev:web": "npm run build:web -- --watch"
   },
   "repository": {
     "type": "git",
@@ -18,9 +21,12 @@
   },
   "homepage": "https://dashrando.github.io/index.html",
   "devDependencies": {
+    "esbuild": "^0.17.11",
+    "esbuild-serve": "^1.0.1",
     "uglify-js": "^3.17.4"
   },
   "dependencies": {
-    "file-saver": "^2.0.5"
+    "file-saver": "^2.0.5",
+    "idb-keyval": "^6.2.0"
   }
 }
