+++ scripts/lib/bps-patch.js
