+++ scripts/lib/dotnet-random.js
